split Timestamp tests off of 19504
